[META] Move helpers file, and add another targetable node
